First stab at browserifying everything
