+++ demo/index.html
@@ -16,7 +16,7 @@
 
         <script src="../shims/mutationobserver-shim.js"></script>
 
-        <script src="../lib/selleckt.js"></script>
+        <script src="../dist/selleckt.js"></script>
 
         <style>
             .demo {
