+++ lib/KEY_CODES.js
@@ -1,0 +1,10 @@
+'use strict';
+
+var KEY_CODES = {
+    DOWN: 40,
+    UP: 38,
+    ENTER: 13,
+    ESC: 27
+};
+
+module.exports = KEY_CODES;
