+++ lib/global-shims/jquery-wrapper.js
@@ -1,0 +1,2 @@
+'use strict';
+module.exports = window.$;
