+++ package.json
@@ -15,10 +15,13 @@
   "scripts": {
     "test": "./testem"
   },
-  "dependencies": {},
   "devDependencies": {
+    "browserify": "^8.1.3",
+    "browserify-shim": "^3.8.2",
     "grunt": "^0.4.5",
+    "grunt-browserify": "^3.3.0",
     "grunt-cli": "^0.1.13",
+    "grunt-contrib-clean": "^0.6.0",
     "grunt-contrib-connect": "^0.9.0",
     "grunt-contrib-jshint": "^0.11.0",
     "grunt-saucelabs": "^8.5.0",
@@ -34,5 +37,15 @@
       "jquery": "jquery/jquery",
       "underscore": "documentcloud/underscore"
     }
+  },
+  "browserify": {
+    "transform": [
+      "browserify-shim"
+    ]
+  },
+  "browserify-shim": {
+    "jquery": "global:$",
+    "underscore": "global:_",
+    "Mustache": "global:Mustache"
   }
 }
