+++ test/SingleSelleckt.tests.js
@@ -1,97 +1,229 @@
-define(['lib/selleckt', 'lib/mustache.js'],
-    function(Selleckt, Mustache){
-    'use strict';
+'use strict';
 
-    describe('selleckt', function(){
-        var selleckt,
-            mainTemplate =
-                '<div class="{{className}}">' +
+var SingleSelleckt = require('../lib/SingleSelleckt');
+
+var $ = require('jquery');
+var _ = require('underscore');
+var Mustache = require('Mustache');
+
+describe('SingleSelleckt', function(){
+    var selleckt,
+        mainTemplate =
+            '<div class="{{className}}">' +
+            '<div class="selected">' +
+                '<span class="selectedText">{{selectedItemText}}</span><i class="icon-arrow-down"></i>' +
+            '</div>' +
+            '<ul class="items">' +
+                '{{#items}}' +
+                '<li class="item{{#selected}} selected{{/selected}}" data-value="{{value}}">' +
+                    '{{label}}' +
+                '</li>' +
+                '{{/items}}' +
+            '</ul>' +
+            '</div>',
+        elHtml =
+            '<select>' +
+                '<option selected value="1">foo</option>' +
+                '<option value="2" data-meh="whee" data-bah="oink">bar</option>' +
+                '<option value="3">baz</option>' +
+            '</select>',
+        $el;
+
+    beforeEach(function(){
+        $el = $(elHtml).appendTo('body');
+    });
+
+    afterEach(function(){
+        $el.remove();
+        $el = undefined;
+
+        if(selleckt){
+            selleckt.destroy();
+            selleckt = undefined;
+        }
+    });
+
+    describe('Instantiation', function(){
+        var template =
+            '<div class="{{className}}" tabindex=1>' +
                 '<div class="selected">' +
                     '<span class="selectedText">{{selectedItemText}}</span><i class="icon-arrow-down"></i>' +
                 '</div>' +
                 '<ul class="items">' +
-                    '{{#items}}' +
-                    '<li class="item{{#selected}} selected{{/selected}}" data-value="{{value}}">' +
-                        '{{label}}' +
+                    '{{#showSearch}}' +
+                    '<li class="searchContainer">' +
+                        '<input class="search"></input>' +
                     '</li>' +
+                    '{{/showSearch}}' +
+                    '{{#items}}' +
+                        '{{> item}}' +
                     '{{/items}}' +
                 '</ul>' +
-                '</div>',
-            elHtml =
-                '<select>' +
-                    '<option selected value="1">foo</option>' +
-                    '<option value="2" data-meh="whee" data-bah="oink">bar</option>' +
-                    '<option value="3">baz</option>' +
-                '</select>',
-            $el;
+            '</div>',
+            itemTemplate = '<li class="{{itemClass}}" data-text="{{label}}" data-value="{{value}}">' +
+                '<span class="{{itemTextClass}}">{{label}}</span>' +
+            '</li>';
 
-        beforeEach(function(){
-            $el = $(elHtml).appendTo('body');
+        describe('invalid instantiation', function(){
+            it('pukes if instantiated with an invalid template format', function(){
+                var err;
+
+                try{
+                    selleckt = new SingleSelleckt({
+                        mainTemplate : {template: template},
+                        $selectEl : $el,
+                        className: 'selleckt',
+                        selectedClass: 'selected',
+                        selectedTextClass: 'selectedText',
+                        itemsClass: 'items',
+                        itemClass: 'item',
+                        selectedClassName: 'isSelected',
+                        highlightClassName: 'isHighlighted'
+                    });
+                } catch(e){
+                    err = e;
+                }
+
+                expect(err).toBeDefined();
+                expect(err.message).toEqual('Please provide a valid mustache template.');
+            });
+
         });
 
-        afterEach(function(){
-            $el.remove();
-            $el = undefined;
+        describe('valid instantation', function(){
 
-            if(selleckt){
-                selleckt.destroy();
-                selleckt = undefined;
-            }
-        });
+            beforeEach(function(){
+                selleckt = new SingleSelleckt({
+                    mainTemplate: template,
+                    itemTemplate: itemTemplate,
+                    $selectEl : $el,
+                    className: 'selleckt',
+                    selectedClass: 'selected',
+                    selectedTextClass: 'selectedText',
+                    itemsClass: 'items',
+                    itemClass: 'item',
+                    selectedClassName: 'isSelected',
+                    highlightClass: 'isHighlighted'
+                });
+            });
 
-        describe('Instantiation', function(){
-            var template =
-                '<div class="{{className}}" tabindex=1>' +
-                    '<div class="selected">' +
-                        '<span class="selectedText">{{selectedItemText}}</span><i class="icon-arrow-down"></i>' +
-                    '</div>' +
-                    '<ul class="items">' +
-                        '{{#showSearch}}' +
-                        '<li class="searchContainer">' +
-                            '<input class="search"></input>' +
-                        '</li>' +
-                        '{{/showSearch}}' +
-                        '{{#items}}' +
-                            '{{> item}}' +
-                        '{{/items}}' +
-                    '</ul>' +
-                '</div>',
-                itemTemplate = '<li class="{{itemClass}}" data-text="{{label}}" data-value="{{value}}">' +
-                    '<span class="{{itemTextClass}}">{{label}}</span>' +
-                '</li>';
+            it('stores the selectedClass as this.selectedClass', function(){
+                expect(selleckt.selectedClass).toEqual('selected');
+            });
 
-            describe('invalid instantiation', function(){
-                it('pukes if instantiated with an invalid template format', function(){
-                    var err;
+            it('stores the selectedTextClass as this.selectedClass', function(){
+                expect(selleckt.selectedTextClass).toEqual('selectedText');
+            });
 
-                    try{
-                        selleckt = Selleckt.create({
-                            mainTemplate : {template: template},
-                            $selectEl : $el,
+            it('stores the itemsClass as this.itemsClass', function(){
+                expect(selleckt.itemsClass).toEqual('items');
+            });
+
+            it('stores the itemClass as this.itemClass', function(){
+                expect(selleckt.itemClass).toEqual('item');
+            });
+
+            it('stores options.mainTemplate as this.template', function(){
+                expect(Mustache.render(selleckt.mainTemplate, {})).toEqual(Mustache.render(template, {}));
+            });
+
+            it('stores options.itemTemplate as this.itemTemplate', function(){
+                expect(Mustache.render(selleckt.itemTemplate, {})).toEqual(Mustache.render(itemTemplate, {}));
+            });
+
+            it('stores options.mainTemplateData as this.mainTemplateData', function(){
+                expect(selleckt.mainTemplateData).toEqual({});
+            });
+
+            it('stores options.selectEl as this.originalSelectEl', function(){
+                expect(selleckt.$originalSelectEl).toEqual($el);
+            });
+
+            it('stores options.className as this.className', function(){
+                expect(selleckt.className).toEqual('selleckt');
+            });
+
+            it('stores options.highlightClass as this.highlightClass', function(){
+                expect(selleckt.highlightClass).toEqual('isHighlighted');
+            });
+
+            describe('items', function(){
+                it('instantiates this.items as an based on the options in the original select', function(){
+                    expect(selleckt.items.length).toEqual(3);
+                });
+                it('stores the option text as "label"', function(){
+                    expect(selleckt.items[0].label).toEqual('foo');
+                    expect(selleckt.items[1].label).toEqual('bar');
+                });
+                it('stores the option value as "value"', function(){
+                    expect(selleckt.items[0].value).toEqual('1');
+                    expect(selleckt.items[1].value).toEqual('2');
+                });
+                it('stores all the option data attributes in "data"', function(){
+                    expect(selleckt.items[0].data).toBeDefined();
+                    expect(_.size(selleckt.items[0].data)).toEqual(0);
+
+                    expect(selleckt.items[1].data).toBeDefined();
+                    expect(_.size(selleckt.items[1].data)).toEqual(2);
+                    expect(selleckt.items[1].data).toEqual({
+                        meh: 'whee',
+                        bah: 'oink'
+                    });
+                });
+                it('stores the selected option as this.selectedItem', function(){
+                    expect(selleckt.selectedItem).toBeDefined();
+                    expect(selleckt.selectedItem).toEqual({
+                        value: '1',
+                        label: 'foo',
+                        data: {}
+                    });
+                });
+
+                describe('when there is no selected item already', function(){
+                    var $newEl;
+
+                    beforeEach(function(){
+                        var selectHtml = '<select>' +
+                        '<option value="1">foo</option>' +
+                        '<option value="2">bar</option>' +
+                        '<option value="3">baz</option>' +
+                        '</select>';
+
+                        selleckt.destroy();
+
+                        $newEl = $(selectHtml).appendTo('body');
+
+                        selleckt = new SingleSelleckt({
+                            mainTemplate: template,
+                            $selectEl : $newEl,
                             className: 'selleckt',
                             selectedClass: 'selected',
                             selectedTextClass: 'selectedText',
                             itemsClass: 'items',
                             itemClass: 'item',
                             selectedClassName: 'isSelected',
-                            highlightClassName: 'isHighlighted'
+                            highlightClass: 'isHighlighted'
                         });
-                    } catch(e){
-                        err = e;
-                    }
+                    });
 
-                    expect(err).toBeDefined();
-                    expect(err.message).toEqual('Please provide a valid mustache template.');
+                    afterEach(function(){
+                        $newEl.remove();
+                        $newEl = undefined;
+                    });
+
+                    it('does not select an item', function(){
+                        expect(selleckt.selectedItem).toBeUndefined();
+                    });
                 });
-
             });
 
-            describe('valid instantation', function(){
+            describe('events', function(){
+                it('does not trigger change event when instantiated on select with selected option', function(){
+                    var onChangeStub = sinon.stub();
+                    $el.on('change', onChangeStub);
 
-                beforeEach(function(){
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate: template,
-                        itemTemplate: itemTemplate,
                         $selectEl : $el,
                         className: 'selleckt',
                         selectedClass: 'selected',
@@ -101,144 +233,15 @@ define(['lib/selleckt', 'lib/mustache.js'],
                         selectedClassName: 'isSelected',
                         highlightClass: 'isHighlighted'
                     });
-                });
 
-                it('stores the selectedClass as this.selectedClass', function(){
-                    expect(selleckt.selectedClass).toEqual('selected');
-                });
-
-                it('stores the selectedTextClass as this.selectedClass', function(){
-                    expect(selleckt.selectedTextClass).toEqual('selectedText');
-                });
-
-                it('stores the itemsClass as this.itemsClass', function(){
-                    expect(selleckt.itemsClass).toEqual('items');
-                });
-
-                it('stores the itemClass as this.itemClass', function(){
-                    expect(selleckt.itemClass).toEqual('item');
-                });
-
-                it('stores options.mainTemplate as this.template', function(){
-                    expect(Mustache.render(selleckt.mainTemplate, {})).toEqual(Mustache.render(template, {}));
-                });
-
-                it('stores options.itemTemplate as this.itemTemplate', function(){
-                    expect(Mustache.render(selleckt.itemTemplate, {})).toEqual(Mustache.render(itemTemplate, {}));
-                });
-
-                it('stores options.mainTemplateData as this.mainTemplateData', function(){
-                    expect(selleckt.mainTemplateData).toEqual({});
-                });
-
-                it('stores options.selectEl as this.originalSelectEl', function(){
-                    expect(selleckt.$originalSelectEl).toEqual($el);
-                });
-
-                it('stores options.className as this.className', function(){
-                    expect(selleckt.className).toEqual('selleckt');
-                });
-
-                it('stores options.highlightClass as this.highlightClass', function(){
-                    expect(selleckt.highlightClass).toEqual('isHighlighted');
-                });
-
-                describe('items', function(){
-                    it('instantiates this.items as an based on the options in the original select', function(){
-                        expect(selleckt.items.length).toEqual(3);
-                    });
-                    it('stores the option text as "label"', function(){
-                        expect(selleckt.items[0].label).toEqual('foo');
-                        expect(selleckt.items[1].label).toEqual('bar');
-                    });
-                    it('stores the option value as "value"', function(){
-                        expect(selleckt.items[0].value).toEqual('1');
-                        expect(selleckt.items[1].value).toEqual('2');
-                    });
-                    it('stores all the option data attributes in "data"', function(){
-                        expect(selleckt.items[0].data).toBeDefined();
-                        expect(_.size(selleckt.items[0].data)).toEqual(0);
-
-                        expect(selleckt.items[1].data).toBeDefined();
-                        expect(_.size(selleckt.items[1].data)).toEqual(2);
-                        expect(selleckt.items[1].data).toEqual({
-                            meh: 'whee',
-                            bah: 'oink'
-                        });
-                    });
-                    it('stores the selected option as this.selectedItem', function(){
-                        expect(selleckt.selectedItem).toBeDefined();
-                        expect(selleckt.selectedItem).toEqual({
-                            value: '1',
-                            label: 'foo',
-                            data: {}
-                        });
-                    });
-
-                    describe('when there is no selected item already', function(){
-                        var $newEl;
-
-                        beforeEach(function(){
-                            var selectHtml = '<select>' +
-                            '<option value="1">foo</option>' +
-                            '<option value="2">bar</option>' +
-                            '<option value="3">baz</option>' +
-                            '</select>';
-
-                            selleckt.destroy();
-
-                            $newEl = $(selectHtml).appendTo('body');
-
-                            selleckt = Selleckt.create({
-                                mainTemplate: template,
-                                $selectEl : $newEl,
-                                className: 'selleckt',
-                                selectedClass: 'selected',
-                                selectedTextClass: 'selectedText',
-                                itemsClass: 'items',
-                                itemClass: 'item',
-                                selectedClassName: 'isSelected',
-                                highlightClass: 'isHighlighted'
-                            });
-                        });
-
-                        afterEach(function(){
-                            $newEl.remove();
-                            $newEl = undefined;
-                        });
-
-                        it('does not select an item', function(){
-                            expect(selleckt.selectedItem).toBeUndefined();
-                        });
-                    });
-                });
-
-                describe('events', function(){
-                    it('does not trigger change event when instantiated on select with selected option', function(){
-                        var onChangeStub = sinon.stub();
-                        $el.on('change', onChangeStub);
-
-                        selleckt = Selleckt.create({
-                            mainTemplate: template,
-                            $selectEl : $el,
-                            className: 'selleckt',
-                            selectedClass: 'selected',
-                            selectedTextClass: 'selectedText',
-                            itemsClass: 'items',
-                            itemClass: 'item',
-                            selectedClassName: 'isSelected',
-                            highlightClass: 'isHighlighted'
-                        });
-
-                        expect(onChangeStub.called).toEqual(false);
-                        $el.off('change', onChangeStub);
-                    });
+                    expect(onChangeStub.called).toEqual(false);
+                    $el.off('change', onChangeStub);
                 });
             });
 
             describe('template formats', function(){
                 it('accepts template strings', function(){
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $el,
                         className: 'selleckt',
@@ -259,7 +262,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 it('can build template data object from custom data and plugin data', function(){
                     var templateData;
 
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         $selectEl : $el,
                         enableSearch: true,
                         className: 'selleckt',
@@ -283,7 +286,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 it('prevents custom data from overriding plugin data', function(){
                     var templateData;
 
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         $selectEl : $el,
                         mainTemplateData: {
                             selectLabel: 'Please selleckt',
@@ -304,7 +307,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 it('includes user configurable class names in template data', function(){
                     var templateData;
 
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         $selectEl: $el,
                         className: 'sellecktTest',
                         selectedClass: 'trigger',
@@ -332,7 +335,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
 
         describe('rendering', function(){
             beforeEach(function(){
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     $selectEl : $el
                 });
 
@@ -373,7 +376,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                     '</div>';
 
                 selleckt.destroy();
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     $selectEl : $el,
                     mainTemplate: customTemplate,
                     mainTemplateData: {
@@ -392,7 +395,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
 
                 $('body').css('overflow-y', 'scroll');
 
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     mainTemplate : mainTemplate,
                     $selectEl : $el
                 });
@@ -415,7 +418,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                     'max-height': '1000px'
                 });
 
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     mainTemplate : mainTemplate,
                     $selectEl : $el
                 });
@@ -444,7 +447,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                     value: 'new value'
                 };
 
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     $selectEl : $el
                 });
 
@@ -729,7 +732,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
             var waitTime = MutationObserver._period ? MutationObserver._period * 2 : 1;
 
             beforeEach(function(){
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     $selectEl : $el
                 });
 
@@ -829,7 +832,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
 
         describe('Events', function(){
             beforeEach(function(){
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     $selectEl : $el
                 });
 
@@ -1214,7 +1217,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 describe('with an empty option', function(){
                     beforeEach(function(){
                         $el = $(elHtml).append('<option></option>').appendTo('body');
-                        selleckt = Selleckt.create({
+                        selleckt = new SingleSelleckt({
                             $selectEl : $el
                         });
 
@@ -1368,7 +1371,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
             describe('initialization', function(){
                 it('displays a searchbox if settings.enableSearch is true and ' +
                         'there are more items than options.searchThreshold', function(){
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml),
                         enableSearch: true,
@@ -1381,7 +1384,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 });
                 it('does not display a searchbox if settings.enableSearch is true and ' +
                         'there are fewer items than options.searchThreshold', function(){
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml),
                         enableSearch: true,
@@ -1393,7 +1396,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                     expect(selleckt.$sellecktEl.find('.searchContainer').length).toEqual(0);
                 });
                 it('does not display a searchbox if settings.enableSearch is false', function(){
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml),
                         enableSearch: false
@@ -1405,7 +1408,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 });
 
                 it('empties the search input when _close() is called', function(){
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml),
                         enableSearch: true
@@ -1425,7 +1428,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 it('clears search when _open() is called', function(){
                     var filterOptionsStub;
 
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml),
                         enableSearch: true
@@ -1444,7 +1447,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
                 it('unfilters the selections list when _close() is called', function(){
                     var filterOptionsStub;
 
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml),
                         enableSearch: true
@@ -1462,10 +1465,8 @@ define(['lib/selleckt', 'lib/mustache.js'],
             });
 
             describe('filtering', function(){
-                var clock;
-
                 beforeEach(function(){
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml),
                         enableSearch: true
@@ -1532,7 +1533,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
 
                 it('escapes the options when filtering and opening', function(){
                     selleckt.destroy();
-                    selleckt = Selleckt.create({
+                    selleckt = new SingleSelleckt({
                         mainTemplate : template,
                         $selectEl : $(selectHtml.replace(/bar/g, '<b>some HTML</b>')),
                         enableSearch: true
@@ -1572,7 +1573,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
 
         describe('removal', function(){
             beforeEach(function(){
-                selleckt = Selleckt.create({
+                selleckt = new SingleSelleckt({
                     mainTemplate : mainTemplate,
                     $selectEl : $el
                 });
@@ -1617,530 +1618,7 @@ define(['lib/selleckt', 'lib/mustache.js'],
 
                 expect(stopObservingMutationsSpy.calledOnce).toEqual(true);
             });
-        });
-    });
 
-    describe('multiselleckt', function(){
-        var multiSelleckt,
-            $el,
-            selectHtml  =
-                '<select multiple>' +
-                    '<option value="1" selected>foo</option>' +
-                    '<option value="2" data-meh="whee" data-bah="oink">bar</option>' +
-                    '<option value="3" selected>baz</option>' +
-                '</select>',
-            mainTemplate =
-                '<div class="{{className}} custom" tabindex=1>' +
-                    '<ul class="mySelections">' +
-                    '{{#selections}}' +
-                    '{{/selections}}' +
-                    '</ul>' +
-                    '<div class="selected">' +
-                        '<span class="selectedText">{{selectedItemText}}</span><i class="icon-arrow-down"></i>' +
-                    '</div>' +
-                    '<ul class="items">' +
-                        '{{#items}}' +
-                        '<li class="item" data-text="{{label}}" data-value="{{value}}">' +
-                            '{{label}}' +
-                        '</li>' +
-                        '{{/items}}' +
-                    '</ul>' +
-                '</div>',
-            selectionTemplate =
-                '<li class="mySelectionItem custom-item" data-value="{{value}}">' +
-                    '{{text}}<i class="icon-remove unselect"></i>' +
-                '</li>';
-
-        beforeEach(function(){
-            $el = $(selectHtml);
-        });
-
-        afterEach(function(){
-            $el = undefined;
-
-            if(multiSelleckt){
-                multiSelleckt.destroy();
-                multiSelleckt = undefined;
-            }
-        });
-
-        describe('initialization', function(){
-            describe('custom options', function(){
-                beforeEach(function(){
-                    multiSelleckt = Selleckt.create({
-                        mainTemplate: mainTemplate,
-                        selectionTemplate: selectionTemplate,
-                        multiple: true,
-                        $selectEl : $el,
-                        className: 'selleckt',
-                        selectedTextClass: 'selectedText',
-                        selectionsClass: 'mySelections',
-                        selectionItemClass: 'mySelectionItem',
-                        placeholderText: 'click me!',
-                        alternatePlaceholder: 'click me again!',
-                        itemsClass: 'items',
-                        itemClass: 'item',
-                        unselectItemClass: 'unselectItem',
-                        selectedClass: 'isSelected',
-                        highlightClass: 'isHighlighted',
-                        showEmptyList: true
-                    });
-                });
-
-                it('stores options.selectionTemplate as this.selectionTemplate',function(){
-                    expect(Mustache.render(multiSelleckt.selectionTemplate, {})).toEqual(Mustache.render(selectionTemplate, {}));
-                });
-                it('stores options.selectionsClass as this.selectionsClass',function(){
-                    expect(multiSelleckt.selectionsClass).toEqual('mySelections');
-                });
-                it('stores options.selectionItemClass as this.selectionItemClass',function(){
-                    expect(multiSelleckt.selectionItemClass).toEqual('mySelectionItem');
-                });
-                it('stores options.placeholderText as this.placeholderText',function(){
-                    expect(multiSelleckt.placeholderText).toEqual('click me!');
-                });
-                it('stores options.alternatePlaceholder as this.alternatePlaceholder',function(){
-                    expect(multiSelleckt.alternatePlaceholder).toEqual('click me again!');
-                });
-                it('stores options.unselectItemClass as this.unselectItemClass',function(){
-                    expect(multiSelleckt.unselectItemClass).toEqual('unselectItem');
-                });
-                it('stores options.selectedClassName as this.selectedClass',function(){
-                    expect(multiSelleckt.selectedClass).toEqual('isSelected');
-                });
-                it('stores options.highlightClass as this.highlightClass',function(){
-                    expect(multiSelleckt.highlightClass).toEqual('isHighlighted');
-                });
-                it('stores options.showEmptyList as this.showEmptyList',function(){
-                    expect(multiSelleckt.showEmptyList).toEqual(true);
-                });
-            });
-
-            describe('defaults', function(){
-                beforeEach(function(){
-                    multiSelleckt = Selleckt.create({
-                        multiple: true,
-                        $selectEl : $el
-                    });
-                });
-
-                it('defaults this.selectionsClass to "selections"',function(){
-                    expect(multiSelleckt.selectionsClass).toEqual('selections');
-                });
-                it('defaults this.selectionItemClass to "selectionItem"',function(){
-                    expect(multiSelleckt.selectionItemClass).toEqual('selectionItem');
-                });
-                it('defaults this.unselectItemClass to "unselect"',function(){
-                    expect(multiSelleckt.unselectItemClass).toEqual('unselect');
-                });
-                it('defaults this.placeholderText to "Please select..."',function(){
-                    expect(multiSelleckt.placeholderText).toEqual('Please select...');
-                });
-                it('defaults this.alternatePlaceholder to "Select another..."',function(){
-                    expect(multiSelleckt.alternatePlaceholder).toEqual('Select another...');
-                });
-                it('defaults this.showEmptyList to false',function(){
-                    expect(multiSelleckt.showEmptyList).toEqual(false);
-                });
-            });
-        });
-
-        describe('template formats', function(){
-            it('accepts template strings', function(){
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl : $el,
-                    selectionsClass: 'mySelections',
-                    selectionItemClass: 'mySelectionItem',
-                    mainTemplate : mainTemplate,
-                    selectionTemplate : selectionTemplate
-                });
-
-                multiSelleckt.render();
-
-                expect(multiSelleckt.$sellecktEl.find('.mySelectionItem').length).toEqual(2);
-                expect(multiSelleckt.$sellecktEl.find('.'+multiSelleckt.selectionItemClass).eq(0).text()).toEqual('foo');
-                expect(multiSelleckt.$sellecktEl.find('.'+multiSelleckt.selectionItemClass).eq(1).text()).toEqual('baz');
-            });
-        });
-
-        describe('template data', function(){
-            it('includes user configurable class name in main template data', function(){
-                var templateData;
-
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl: $el,
-                    selectionsClass: 'selected-list'
-                });
-
-                templateData = multiSelleckt.getTemplateData();
-
-                expect(templateData.selectionsClass).toEqual('selected-list');
-            });
-
-            it('includes user configurable class names in item template data', function(){
-                var itemData;
-
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl: $el,
-                    selectionItemClass: 'selected-item',
-                    unselectItemClass: 'unselect'
-                });
-
-                itemData = multiSelleckt.getItemTemplateData(multiSelleckt.items[0]);
-
-                expect(itemData.selectionItemClass).toEqual('selected-item');
-                expect(itemData.unselectItemClass).toEqual('unselect');
-            });
-
-            it('includes option data attributes as property of item template data', function(){
-                var itemData;
-
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl: $el
-                });
-
-                itemData = multiSelleckt.getItemTemplateData(multiSelleckt.items[1]);
-
-                expect(itemData.data).toBeDefined();
-                expect(itemData.data.meh).toEqual('whee');
-                expect(itemData.data.bah).toEqual('oink');
-            });
-        });
-
-        describe('rendering', function(){
-            beforeEach(function(){
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl : $el
-                });
-
-                multiSelleckt.render();
-            });
-
-            it('renders the selected items in the multiselectItemTemplate', function(){
-                expect(multiSelleckt.getSelection()).toEqual([{
-                        value: '1',
-                        label: 'foo',
-                        data: {}
-                    }, {
-                        value: '3',
-                        label: 'baz',
-                        data: {}
-                    }]);
-
-                expect(multiSelleckt.$sellecktEl.find('.selectionItem').length).toEqual(2);
-                expect(multiSelleckt.$sellecktEl.find('.'+multiSelleckt.selectionItemClass).eq(0).text()).toEqual('foo');
-                expect(multiSelleckt.$sellecktEl.find('.'+multiSelleckt.selectionItemClass).eq(1).text()).toEqual('baz');
-            });
-            it('attaches the item to the selectedItem dom element', function(){
-                var selectedItems = multiSelleckt.getSelection();
-
-                expect(selectedItems).toEqual([{
-                        value: '1',
-                        label: 'foo',
-                        data: {}
-                    }, {
-                        value: '3',
-                        label: 'baz',
-                        data: {}
-                    }]);
-
-                expect(multiSelleckt.$sellecktEl.find('.selectionItem').eq(0).data('item')).toEqual(selectedItems[0]);
-                expect(multiSelleckt.$sellecktEl.find('.selectionItem').eq(1).data('item')).toEqual(selectedItems[1]);
-            });
-        });
-
-        describe('item selection', function(){
-            beforeEach(function(){
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl : $el
-                });
-            });
-
-            it('allows multiple items to be selected', function(){
-                multiSelleckt.render();
-                multiSelleckt.selectedItems = [];
-
-                multiSelleckt.$sellecktEl.find('li.item').eq(0).trigger('mouseout');
-                multiSelleckt.$sellecktEl.find('li.item').eq(1).trigger('mouseover').trigger('click');
-
-                expect(multiSelleckt.getSelection()).toEqual([{
-                    value: '2',
-                    label: 'bar',
-                    data: {
-                        meh: 'whee',
-                        bah: 'oink'
-                    }
-                }]);
-
-                multiSelleckt.$sellecktEl.find('li.item').eq(1).trigger('mouseout');
-                multiSelleckt.$sellecktEl.find('li.item').eq(0).trigger('mouseover').trigger('click');
-
-                expect(multiSelleckt.getSelection()).toEqual([
-                    {
-                        value: '2',
-                        label: 'bar',
-                        data: {
-                            meh: 'whee',
-                            bah: 'oink'
-                        }
-                    }, {
-                        value: '1',
-                        label: 'foo',
-                        data: {}
-                    }
-                ]);
-            });
-
-            it('hides the selected item from the list', function(){
-                multiSelleckt.selectedItems = [];
-                multiSelleckt.render();
-
-                multiSelleckt.$sellecktEl.find('li.item').eq(0).trigger('mouseout');
-                multiSelleckt.$sellecktEl.find('li.item').eq(1).trigger('mouseover').trigger('click');
-
-                expect(multiSelleckt.getSelection()).toEqual([{
-                    value: '2',
-                    label: 'bar',
-                    data: {
-                        meh: 'whee',
-                        bah: 'oink'
-                    }
-                }]);
-
-                multiSelleckt._open();
-
-                expect(multiSelleckt.$items.find('.item[data-value="2"]').css('display')).toEqual('none');
-            });
-
-            it('adds a class of "disabled" to the select if all options are selected', function(){
-                multiSelleckt.render();
-
-                multiSelleckt.$sellecktEl.find('li.item').eq(1).trigger('mouseover').trigger('click');
-
-                expect(multiSelleckt.getSelection().length).toEqual(3);
-
-                expect(multiSelleckt.$sellecktEl.hasClass('disabled')).toEqual(true);
-            });
-
-            it('does not add a class of "disabled" to the select if all options are selected but options.showEmptyList is true', function(){
-                multiSelleckt.destroy();
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl : $el,
-                    showEmptyList: true
-                });
-                multiSelleckt.render();
-
-                multiSelleckt.$sellecktEl.find('li.item').eq(1).trigger('mouseover').trigger('click');
-
-                expect(multiSelleckt.getSelection().length).toEqual(3);
-
-                expect(multiSelleckt.$sellecktEl.hasClass('disabled')).toEqual(false);
-            });
-
-            it('updates the original select element with the new value', function(){
-                multiSelleckt.selectedItems = [];
-                multiSelleckt.render();
-
-                multiSelleckt.selectItem(multiSelleckt.items[1]);
-
-                expect(multiSelleckt.$originalSelectEl.val()).toEqual([multiSelleckt.items[1].value]);
-
-                multiSelleckt.selectItem(multiSelleckt.items[2]);
-                expect(multiSelleckt.$originalSelectEl.val()).toEqual([multiSelleckt.items[1].value, multiSelleckt.items[2].value]);
-            });
-
-            describe('item deselection', function(){
-                it('removes an item when the "unselect" link is clicked', function(){
-                    multiSelleckt.render();
-
-                    expect(multiSelleckt.getSelection().length).toEqual(2);
-
-                    expect(multiSelleckt.$sellecktEl.find('.selectionItem').length).toEqual(2);
-
-                    multiSelleckt.$sellecktEl.find('.selectionItem .unselect').first().trigger('click');
-
-                    expect(multiSelleckt.$sellecktEl.find('.selectionItem').length).toEqual(1);
-                });
-                it('removes the class of "disabled" from the select if all options were selected but one becomes available', function(){
-                    multiSelleckt.render();
-
-                    multiSelleckt.$sellecktEl.find('li.item').eq(1).trigger('mouseover').trigger('click');
-
-                    expect(multiSelleckt.getSelection().length).toEqual(3);
-                    expect(multiSelleckt.$sellecktEl.hasClass('disabled')).toEqual(true);
-
-                    multiSelleckt.$sellecktEl.find('.selectionItem .unselect').first().trigger('click');
-
-                    expect(multiSelleckt.getSelection().length).toEqual(2);
-                    expect(multiSelleckt.$sellecktEl.hasClass('disabled')).toEqual(false);
-                });
-                it('updates the original select element with the removed value', function(){
-                    multiSelleckt.selectedItems = [];
-                    multiSelleckt.render();
-
-                    multiSelleckt.selectItem(multiSelleckt.items[1]);
-
-                    expect(multiSelleckt.$originalSelectEl.val()).toEqual([multiSelleckt.items[1].value]);
-
-                    multiSelleckt.selectItem(multiSelleckt.items[2]);
-                    expect(multiSelleckt.$originalSelectEl.val()).toEqual([multiSelleckt.items[1].value, multiSelleckt.items[2].value]);
-
-                    multiSelleckt.unselectItem(multiSelleckt.items[2]);
-
-                    expect(multiSelleckt.$originalSelectEl.val()).toEqual([multiSelleckt.items[1].value]);
-                });
-                it('restores the placeholder text when the selection is cleared', function(){
-                    multiSelleckt.render();
-                    multiSelleckt.selectItem(multiSelleckt.items[1]);
-
-                    expect(multiSelleckt.selectedItems.length).toEqual(3);
-                    expect(multiSelleckt.$sellecktEl.find('.'+multiSelleckt.selectedTextClass).text())
-                        .toEqual(multiSelleckt.alternatePlaceholder);
-
-                    multiSelleckt.unselectItem(multiSelleckt.items[0]);
-                    multiSelleckt.unselectItem(multiSelleckt.items[1]);
-                    multiSelleckt.unselectItem(multiSelleckt.items[2]);
-
-                    expect(multiSelleckt.selectedItems.length).toEqual(0);
-                    expect(multiSelleckt.$sellecktEl.find('.'+multiSelleckt.selectedTextClass).text())
-                        .toEqual(multiSelleckt.placeholderText);
-                });
-            });
-        });
-
-        describe('unselecting items', function(){
-            var $clickTarget;
-
-            beforeEach(function(){
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl : $el
-                });
-
-                multiSelleckt.render();
-
-                expect(multiSelleckt.getSelection()).toEqual([{
-                    value: '1',
-                    label: 'foo',
-                    data: {}
-                }, {
-                    value: '3',
-                    label: 'baz',
-                    data: {}
-                }]);
-
-                $clickTarget = multiSelleckt.$sellecktEl.find('.'+multiSelleckt.unselectItemClass).eq(0);
-            });
-            it('removes the item from the selections when the unselectItem link is clicked', function(){
-                var $selections = multiSelleckt.$selections;
-
-                expect($selections.children().length).toEqual(2);
-
-                $clickTarget.trigger('click');
-
-                expect($selections.children().length).toEqual(1);
-                expect($selections.children().first().data('item')).toEqual({
-                    value: '3',
-                    label: 'baz',
-                    data: {}
-                });
-            });
-            it('removes the item from the selectedItems array when the remove link is clicked', function(){
-               var $selections = multiSelleckt.$selections;
-
-                expect(multiSelleckt.getSelection()).toEqual([{
-                    value: '1',
-                    label: 'foo',
-                    data: {}
-                }, {
-                    value: '3',
-                    label: 'baz',
-                    data: {}
-                }]);
-
-                $clickTarget.trigger('click');
-
-                expect($selections.children().length).toEqual(1);
-
-                expect(multiSelleckt.getSelection()).toEqual([{
-                    value: '3',
-                    label: 'baz',
-                    data: {}
-                }]);
-            });
-            it('shows the item in the items when the remove link is clicked', function(){
-                var $selections = multiSelleckt.$selections;
-
-                expect(multiSelleckt.$sellecktEl.find('.item[data-value="1"]').css('display')).toEqual('none');
-
-                $clickTarget.trigger('click');
-
-                expect($selections.children().length).toEqual(1);
-                expect(multiSelleckt.$sellecktEl.find('.item[data-value="1"]').css('display')).not.toEqual('none');
-            });
-            it('triggers a change event on original select when item is removed', function(){
-                var changeHandler = sinon.spy();
-
-                multiSelleckt.$originalSelectEl.on('change', changeHandler);
-                $clickTarget.trigger('click');
-
-                expect(changeHandler.calledOnce).toEqual(true);
-                expect(changeHandler.args[0].length).toEqual(2);
-                expect(changeHandler.args[0][1].origin).toEqual('selleckt');
-
-                multiSelleckt.$originalSelectEl.off('change', changeHandler);
-            });
-            it('triggers an "itemUnselected" event with the removed item', function(){
-                var spy = sinon.spy();
-
-                multiSelleckt.bind('itemUnselected', spy);
-                $clickTarget.trigger('click');
-
-                expect(spy.calledOnce).toEqual(true);
-                expect(spy.args[0][0]).toEqual({
-                    value: '1',
-                    label: 'foo',
-                    data: {}
-                });
-            });
-        });
-
-        describe('events', function(){
-            beforeEach(function(){
-                multiSelleckt = Selleckt.create({
-                    multiple: true,
-                    $selectEl : $el
-                });
-            });
-
-            it('does not open when it has a class of "disabled"', function(){
-                multiSelleckt.render();
-
-                multiSelleckt.$sellecktEl.find('li.item').eq(1).trigger('mouseover').trigger('click');
-
-                expect(multiSelleckt.getSelection().length).toEqual(3);
-                expect(multiSelleckt.$sellecktEl.hasClass('disabled')).toEqual(true);
-
-                multiSelleckt.$sellecktEl.find('.'+multiSelleckt.selectedClass).trigger('click');
-
-                expect(multiSelleckt.$sellecktEl.hasClass('open')).toEqual(false);
-            });
-            it('updates multiSelleckt when change is triggered on original select', function(){
-                multiSelleckt.render();
-
-                multiSelleckt.$originalSelectEl.val(['1', '2']).change();
-
-                expect(multiSelleckt.getSelection().length).toEqual(2);
-                expect(multiSelleckt.getSelection()[0].value).toEqual('1');
-                expect(multiSelleckt.getSelection()[1].value).toEqual('2');
-            });
         });
     });
 });
