+++ test/index.html
@@ -13,10 +13,13 @@
     <script src="../shims/es5-shim.min.js"></script>
     <![endif]-->
 
+    <script src="lib/jquery-1.10.1.js"></script>
+    <script src="lib/underscore.js"></script>
+    <script src="lib/mustache.js"></script>
+
     <script src="lib/mocha/mocha.js"></script>
     <script src="lib/expectations.js"></script>
     <script src="lib/sinon-1.1.0.js"></script>
-    <script src="../shims/mutationobserver-shim.js"></script>
     <script src="/testem.js"></script>
 
     <style>
@@ -40,19 +43,46 @@
     <div id="mocha"></div>
 
     <script>
-        var require = {
-            baseUrl: '../',
-            paths: {
-                'jquery': 'test/lib/jquery-1.10.1',
-                'underscore': 'test/lib/underscore',
-                'mustache': 'test/lib/mustache'
-            },
-            shim: {
-                'underscore': {
-                    exports: '_'
-                }
+    mocha.setup({
+        ui: 'bdd'
+    });
+    </script>
+
+    <script src="browserified_tests.js"></script>
+
+    <script>
+    'use strict';
+    var runner = mocha.run();
+    var failedTests = [];
+
+    function logFailure(test, err){
+        var flattenTitles = function(test){
+            var titles = [];
+            while (test.parent.title){
+                titles.push(test.parent.title);
+                test = test.parent;
             }
+            return titles.reverse();
         };
+
+        failedTests.push({
+            name: test.title,
+            result: false,
+            message: err.message,
+            stack: err.stack,
+            titles: flattenTitles(test)
+        });
+    }
+
+    runner.on('end', function(){
+        window.mochaResults = runner.stats;
+        window.mochaResults.reports = failedTests;
+    });
+
+    runner.on('fail', logFailure);
+    </script>
+
+    <script>
         if(!window.console){
             window.console = {
                 log: function(){},
@@ -64,10 +94,6 @@
         window.console.time = window.console.time || function(){};
         window.console.timeEnd = window.console.timeEnd || function(){};
     </script>
-    <script src="lib/require.js"></script>
-
-    <!-- include spec files here... -->
-    <script src="loader.js"></script>
 
 </body>
 </html>
