+++ test/suite.js
@@ -1,0 +1,2 @@
+require('./SingleSelleckt.tests.js');
+require('./MultiSelleckt.tests.js');
